refactor(storybook-info): small refactorings
